# lms-openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/resources": {
            "post": {
                "operationId": "ResourcesController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateResourceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Resource created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Add a new resource to a session (URL, GitHub, or pre-uploaded file)",
                "tags": [
                    "Resources"
                ]
            }
        },
        "/resources/upload": {
            "post": {
                "operationId": "ResourcesController_uploadFile",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "sessionId": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "PDF",
                                            "ZIP"
                                        ]
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "sessionId",
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded and resource created"
                    },
                    "400": {
                        "description": "No file provided or invalid input"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Upload and attach a file resource (PDF, ZIP) to a session",
                "tags": [
                    "Resources"
                ]
            }
        },
        "/resources/reorder": {
            "post": {
                "operationId": "ResourcesController_reorder",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReorderResourcesDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resources reordered successfully"
                    },
                    "400": {
                        "description": "Invalid resource IDs supplied"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Reorder resources within a session",
                "tags": [
                    "Resources"
                ]
            }
        },
        "/resources/session/{sessionId}": {
            "get": {
                "operationId": "ResourcesController_findAllBySession",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "path",
                        "description": "Session UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resources fetched successfully"
                    },
                    "404": {
                        "description": "Session not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get all resources for a session",
                "tags": [
                    "Resources"
                ]
            }
        },
        "/resources/{id}": {
            "get": {
                "operationId": "ResourcesController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Resource UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource fetched successfully"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get a resource by ID",
                "tags": [
                    "Resources"
                ]
            },
            "patch": {
                "operationId": "ResourcesController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Resource UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateResourceDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resource updated successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update a resource",
                "tags": [
                    "Resources"
                ]
            },
            "delete": {
                "operationId": "ResourcesController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Resource UUID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource deleted successfully"
                    },
                    "403": {
                        "description": "Forbidden: Author mismatch"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete a resource",
                "tags": [
                    "Resources"
                ]
            }
        },
...
    },
    "info": {
        "title": "VirtuaLMS API",
        "description": "The VirtuaLMS API description",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateResourceDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "example": "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1234",
                        "description": "Session UUID to attach this resource to"
                    },
                    "title": {
                        "type": "string",
                        "example": "Lesson Cheatsheet & Slides",
                        "description": "Title of the resource"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "PDF",
                            "ZIP",
                            "EXTERNAL_LINK",
                            "GITHUB"
                        ],
                        "example": "PDF",
                        "description": "Type of resource (PDF, ZIP, EXTERNAL_LINK, GITHUB)"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://github.com/example/course-repo",
                        "description": "External link or download URL for the resource"
                    },
                    "storageKey": {
                        "type": "string",
                        "example": "resources/1727680000_cheatsheet.pdf",
                        "description": "Storage key on server / S3 storage for downloadable files"
                    },
                    "fileName": {
                        "type": "string",
                        "example": "cheatsheet.pdf",
                        "description": "Original file name"
                    },
                    "fileSize": {
                        "type": "number",
                        "example": 1048576,
                        "description": "File size in bytes"
                    },
                    "mimeType": {
                        "type": "string",
                        "example": "application/pdf",
                        "description": "MIME type of the resource file"
                    },
                    "position": {
                        "type": "number",
                        "example": 0,
                        "description": "Position index within the session (auto-assigned if omitted)"
                    }
                },
                "required": [
                    "sessionId",
                    "title",
                    "type"
                ]
            },
            "ReorderResourcesDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "example": "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1234",
                        "description": "Session UUID whose resources are being reordered"
                    },
                    "resourceIds": {
                        "example": [
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d1111",
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d2222",
                            "2b7c3f7a-2d2d-4f3b-9e7f-7b6a9c5d3333"
                        ],
                        "description": "Array of Resource IDs in the desired order (index will become the position)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "sessionId",
                    "resourceIds"
                ]
            },
            "UpdateResourceDto": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Updated Resource Title",
                        "description": "Title of the resource"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "PDF",
                            "ZIP",
                            "EXTERNAL_LINK",
                            "GITHUB"
                        ],
                        "description": "Type of resource (PDF, ZIP, EXTERNAL_LINK, GITHUB)"
                    },
                    "url": {
                        "type": "string",
                        "example": "https://github.com/example/updated-course-repo",
                        "description": "External link or download URL for the resource"
                    },
                    "storageKey": {
                        "type": "string",
                        "example": "resources/1727680000_updated_cheatsheet.pdf",
                        "description": "Storage key on server / S3 storage"
                    },
                    "fileName": {
                        "type": "string",
                        "example": "updated_cheatsheet.pdf",
                        "description": "Original file name"
                    },
                    "fileSize": {
                        "type": "number",
                        "example": 2097152,
                        "description": "File size in bytes"
                    },
                    "mimeType": {
                        "type": "string",
                        "example": "application/pdf",
                        "description": "MIME type of the resource file"
                    },
                    "position": {
                        "type": "number",
                        "example": 1,
                        "description": "New position index within the session"
                    }
                }
            },
...
        }
    }
}